+++ src/ResultSetTest.php
@@ -15,7 +15,6 @@ namespace League\Csv;
 
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\Group;
-use PHPUnit\Framework\TestCase;
 use SplTempFileObject;
 
 use function current;
@@ -24,7 +23,7 @@ use function json_encode;
 use function next;
 
 #[Group('reader')]
-final class ResultSetTest extends TestCase
+final class ResultSetTest extends TabularDataReaderTestCase
 {
     private Reader $csv;
     private Statement $stmt;
@@ -47,6 +46,31 @@ final class ResultSetTest extends TestCase
     protected function tearDown(): void
     {
         unset($this->csv, $this->stmt);
+    }
+
+    protected function tabularData(): TabularDataReader
+    {
+        return new ResultSet([
+            ['date', 'temperature', 'place'],
+            ['2011-01-01', '1', 'Galway'],
+            ['2011-01-02', '-1', 'Galway'],
+            ['2011-01-03', '0', 'Galway'],
+            ['2011-01-01', '6', 'Berkeley'],
+            ['2011-01-02', '8', 'Berkeley'],
+            ['2011-01-03', '5', 'Berkeley'],
+        ]);
+    }
+
+    protected function tabularDataWithHeader(): TabularDataReader
+    {
+        return new ResultSet([
+            ['2011-01-01', '1', 'Galway'],
+            ['2011-01-02', '-1', 'Galway'],
+            ['2011-01-03', '0', 'Galway'],
+            ['2011-01-01', '6', 'Berkeley'],
+            ['2011-01-02', '8', 'Berkeley'],
+            ['2011-01-03', '5', 'Berkeley'],
+        ], ['date', 'temperature', 'place']);
     }
 
     public function testFilter(): void